Implementing global variables

+++ src/chunk.c
@@ -73,7 +73,7 @@ uint32_t getLine(const Chunk* chunk, const size_t instructionOffset) {
   }
 }
 
-size_t addConstant(Chunk* chunk, const Value value) {
+uint32_t addConstant(Chunk* chunk, const Value value) {
   writeValueArray(&chunk->constants, value);
 
   return chunk->constants.count - 1;
